+++ admin/pages/delete/product.php
@@ -1,21 +1,25 @@
 <?php
-include "../../config/config.php";
-$id = mysqli_real_escape_string($conn, $_GET['id']);
 
-$sql = "select * from product_image where id='$id'";
-$result = $conn->query($sql)->fetch_all(MYSQLI_ASSOC);
+if (isset($_GET['id'])) {
+
+    include "../../config/config.php";
+    $id = mysqli_real_escape_string($conn, $_GET['id']);
+
+    $sql = "select * from product_image where id='$id'";
+    $result = $conn->query($sql)->fetch_all(MYSQLI_ASSOC);
 
 
-$sql = "delete from product_image where id='$id'";
-if ($conn->query($sql)) {
-
-    $sql = "delete from products where id='$id'";
+    $sql = "delete from product_image where id='$id'";
     if ($conn->query($sql)) {
-        $_SESSION['product_deleted'] = "successful";
-        foreach ($result as $key) {
-            unlink(product_upload . $key['name']);
-        }
-    } else
-        $_SESSION['product_deleted'] = "error";
+
+        $sql = "delete from products where id='$id'";
+        if ($conn->query($sql)) {
+            $_SESSION['product_deleted'] = "successful";
+            foreach ($result as $key) {
+                unlink(product_upload . $key['name']);
+            }
+        } else
+            $_SESSION['product_deleted'] = "error";
+    }
 }
 header("Location:" . url . "products.php");
